requirements.txt
numpy>=1.24.0
scikit-learn>=1.2.0
dtw-python>=1.3.0
torch>=2.0.0
tigramite>=5.1.0
python-dotenv>=1.0.0